refactor(backend): split database.py into db/ package
Extract DatabaseContext, StatusRepository, DeviceLogRepository,
OutageRepository, and MetadataRepository into separate files under
backend/db/. The original backend/database.py becomes a re-export
shim so all existing imports continue to work unchanged.
