+++ backend/db/__init__.py
@@ -1,0 +1,15 @@
+"""Database package — re-exports all public names for convenient imports."""
+
+from .context import DatabaseContext
+from .device_log_repo import DeviceLogRepository
+from .metadata_repo import MetadataRepository
+from .outage_repo import OutageRepository
+from .status_repo import StatusRepository
+
+__all__ = [
+    "DatabaseContext",
+    "DeviceLogRepository",
+    "MetadataRepository",
+    "OutageRepository",
+    "StatusRepository",
+]
